Review architecture
