+++ documentationDEV.md
@@ -1,0 +1,12 @@
+La structure a été modifiée de telle sorte que l'on passe des dossiers ejs, css et js à une structure telle que :
+
+• annexe
+• route
+• view
+• test
+
+Le dossier annexe fait office de "Data access layer", contenant tous les ateliers, ainsi que leurs données en mémoire (nom, description).
+
+Le dossier route fait office de "Controller", régissant toutes les routes de l'application.
+
+Le dossier view fait office de "Service layer", contenant le coeur de l'application.
